test: wire chaos suite via trait hooks for Pest closure compatibility
